msm: ipa: fix incorrect pipe number comparison
In ipa_tag_aggr_force_close, pipe number -1 as input indicates that
aggregation force close needs to be applied on all of the pipes.
Due to a comparison between an int and and unsigned int, the option
of -1 is not being exercised.
Similar instances which share this comparison were fixed as well.

Acked-by: David Arinzon <darinzon@qti.qualcomm.com>

+++ drivers/platform/msm/ipa/ipa_utils.c
@@ -4622,7 +4622,7 @@ int ipa_tag_aggr_force_close(int pipe_num)
 	int num_descs;
 	int num_aggr_descs;
 
-	if (pipe_num < -1 || pipe_num >= ipa_ctx->ipa_num_pipes) {
+	if (pipe_num < -1 || pipe_num >= (int)ipa_ctx->ipa_num_pipes) {
 		IPAERR("Invalid pipe number %d\n", pipe_num);
 		return -EINVAL;
 	}
